requirements.txt
pyserial
RPi.GPIO
picamera2
ultralytics
